fix(coco_preparator_api): change type from uint8 to int32 for alpha mask

+++ pose_estimation/data_preparation/coco_preparator_api.py
@@ -169,10 +169,10 @@ class CocoPreparator:
             img_obj = self._coco.loadImgs(ids_img[i])[0]
             image = io.imread(os.path.join(self._image_folder_path, img_obj['file_name']))
             if img_obj.get('alpha_mask') is not None:
-                alpha_mask = io.imread(os.path.join(self._image_folder_path, img_obj['alpha_mask']))
+                alpha_mask = io.imread(os.path.join(self._image_folder_path, img_obj['alpha_mask'])).astype(np.int32)
             else:
                 # For this image - everything are HUMAN like, i.e. nothing will be changed
-                alpha_mask = np.ones_like(image, dtype=np.uint8) * 255
+                alpha_mask = np.ones_like(image, dtype=np.int32) * 255
             annIds = self._coco.getAnnIds(imgIds=img_obj['id'], iscrowd=None)
             anns = self._coco.loadAnns(annIds)
 
@@ -218,7 +218,7 @@ class CocoPreparator:
             image_tensors.append(image.astype(np.float32, copy=False))
             image_masks.append(image_mask)
             image_properties_tensors.append(np.array(image.shape).astype(np.float32, copy=False))
-            alpha_mask_tensors.append(alpha_mask.astype(np.uint8, copy=False))
+            alpha_mask_tensors.append(alpha_mask.astype(np.int32, copy=False))
             counter += 1
 
             # For large datasets, it's better to save them by parties
@@ -287,10 +287,10 @@ class CocoPreparator:
             img_obj = self._coco.loadImgs(ids_img[i])[0]
             image = io.imread(os.path.join(self._image_folder_path, img_obj['file_name']))
             if img_obj.get('alpha_mask') is not None:
-                alpha_mask = io.imread(os.path.join(self._image_folder_path, img_obj['alpha_mask']))
+                alpha_mask = io.imread(os.path.join(self._image_folder_path, img_obj['alpha_mask'])).astype(np.int32)
             else:
                 # For this image - everything are HUMAN like, i.e. nothing will be changed
-                alpha_mask = np.ones_like(image, dtype=np.uint8) * 255
+                alpha_mask = np.ones_like(image, dtype=np.int32) * 255
             annIds = self._coco.getAnnIds(imgIds=img_obj['id'], iscrowd=None)
             anns = self._coco.loadAnns(annIds)
 
@@ -387,7 +387,7 @@ class CocoPreparator:
             image_tensors.append(image.astype(np.float32, copy=False))
             image_masks.append(image_mask)
             image_properties_tensors.append(np.array(image.shape).astype(np.float32, copy=False))
-            alpha_mask_tensors.append(alpha_mask.astype(np.uint8, copy=False))
+            alpha_mask_tensors.append(alpha_mask.astype(np.int32, copy=False))
 
             counter += 1
 
@@ -464,7 +464,7 @@ class CocoPreparator:
             padding_mask = np.ones((new_h, self._min_image_size, 1)).astype(np.float32, copy=False)
             padding_mask[:, :new_w] = image_mask
 
-            padding_alpha_mask = np.ones((new_h, self._min_image_size, 3)).astype(np.uint8, copy=False)
+            padding_alpha_mask = np.ones((new_h, self._min_image_size, 3)).astype(np.int32, copy=False)
             padding_alpha_mask *= np.min(alpha_mask)
             padding_alpha_mask[:, :new_w] = alpha_mask
             return padding_image, keypoints, padding_mask, padding_alpha_mask
